Trait ncollide2df32​::broad​::Dispatcher​

pub trait Dispatcher<G1, G2, NF> {
    fn dispatch(&self, &G1, &G2) -> Option<NF>;
    fn is_valid(&self, &G1, &G2) -> bool;
}

Trait of dispatcher.

A (collision) dispatcher typically takes two bodies in parameter and return the corresponding narrow phase algorithm.

Required Methods

fn dispatch(&self, &G1, &G2) -> Option<NF>

Deduce the narrow phase from two bodies.

fn is_valid(&self, &G1, &G2) -> bool

Tells whether a collision between two bodies can occur.