Trait ncollide2df32​::broad​::InterferencesBroadPhase​

pub trait InterferencesBroadPhase<B, DV>: BroadPhase<B> {
    fn activate(&mut self, body: &B, f: |&B, &B, &mut DV|);
    fn deactivate(&mut self, &B);
    fn for_each_pair(&self, f: |&B, &B, &DV|);
    fn for_each_pair_mut(&mut self, f: |&B, &B, &mut DV|);
}

Thait of broad phases which check for pairwise interferences.

This is the most common feature provided by broad phases.

Required Methods

fn activate(&mut self, body: &B, f: |&B, &B, &mut DV|)

Marks and object as active.

Active objects are checked for interferences at each update.

fn deactivate(&mut self, &B)

Marks and object as inactive.

Inactive objects are assumed to be static and not tested for mutual interferences.

fn for_each_pair(&self, f: |&B, &B, &DV|)

Execute a function on each interference detected by the broad phase.

fn for_each_pair_mut(&mut self, f: |&B, &B, &mut DV|)

Execute a function on each interference detected by the broad phase.