Struct ncollide2df32​::data​::vec_slice​::VecSliceMut​

pub struct VecSliceMut<'a, T> {
    // some fields omitted
}

A mutable vector slice with a specific length and stride.

The stride is the number of index increments between two elements.

Methods

impl<'a, T> VecSliceMut<'a, T>

fn new(data: &'a mut [T], length: uint, stride: uint) -> VecSliceMut<'a, T>

Creates a new mutable slice.

unsafe fn new_unsafe(data: &'a mut [T], length: uint, stride: uint) -> VecSliceMut<'a, T>

Creates a new mutable slice. The size of the data buffer is not checked.

fn as_slice<'b>(&'b self) -> VecSlice<'b, T>

Creates an immutable slice from this mutable slice.

fn get<'b>(&'b self, i: uint) -> &'b T

Gets the i-th element of the slice.

This is the same as the i * self.stride-th element of the wrapped vector.

fn get_mut<'b>(&'b mut self, i: uint) -> &'b mut T

Gets a mutable reference to the i-th element of the slice without bound-checking.

This is the same as the i * self.stride-th element of the wrapped vector.

unsafe fn unsafe_get<'b>(&'b self, i: uint) -> &'b T

Gets the i-th element of the slice without bound-checking.

This is the same as the i * self.stride-th element of the wrapped vector.

unsafe fn unsafe_get_mut<'b>(&'b mut self, i: uint) -> &'b mut T

Gets a mutable reference to the i-th element of the slice without bound-checking.

This is the same as the i * self.stride-th element of the wrapped vector.

impl<'a, T: Clone> VecSliceMut<'a, T>

fn copy_from(&mut self, data: &VecSlice<T>)

Copy the content of another slice. Both slices must have the same length.