Struct ncollide2df32​::geom​::BezierCurve​

pub struct BezierCurve {
    // some fields omitted
}

Procedural generator of non-rational Bézier curve.

Methods

impl BezierCurve

fn new(control_points: Vec<Vect>) -> BezierCurve

Creates a new non-rational bézier curve.

Parameters:

control_points: The control points of the bézier curve.

fn new_with_degree(degree: uint) -> BezierCurve

Creates a new bézier curve of degree degree.

Its control points are initialized to zero.

fn new_evaluation_cache() -> BezierCurveEvaluationCache

Creates a cache to avaluate a bezier curve.

fn control_points<'a>(&'a self) -> &'a [Vect]

The control points of this bézier curve.

fn control_points_mut<'a>(&'a mut self) -> &'a mut [Vect]

The control points of this bézier curve.

fn npoints(&self) -> uint

The number of control points per u-iso-curve.

fn degree(&self) -> uint

The degree of this curve.

fn at(&self, t: &Scalar, cache: &mut BezierCurveEvaluationCache) -> Vect

Evaluates this bezier curve at the given parameter t.

fn subdivide_at(&self, t: &Scalar, out1: &mut BezierCurve, out2: &mut BezierCurve)

Subdivides this bezier at the given parameter.

The union of the two resulting curves equals the original one.

fn diff(&self, out: &mut BezierCurve)

Computes the derivative of this bezier curve.