Struct ncollide2df32​::geom​::BezierSurfaceEvaluationCache​

pub struct BezierSurfaceEvaluationCache {
    // some fields omitted
}

Cache used to evaluate a bezier surface at a given parameter.

The same cache can be used for multiple evaluations (that is the whole point of this).