Struct ncollide2df32​::geom​::Cuboid​

pub struct Cuboid {
    // some fields omitted
}

Geometry of a box.

Parameters:

Scalar - type of an extent of the box

Vect - vector of extents. This determines the box dimension

Methods

impl Cuboid

fn new(half_extents: Vect) -> Cuboid

Creates a new box from its half-extents. Half-extents are the box half-width along each axis. Each half-extent must be greater than 0.04.

fn new_with_margin(half_extents: Vect, margin: Scalar) -> Cuboid

Creates a new box from its half-extents and its margin. Half-extents are the box half-width along each axis. Each half-extent must be greater than the margin.

impl Cuboid

fn half_extents(&self) -> Vect

The half-extents of this box. Half-extents are the box half-width along each axis.

fn margin(&self) -> Scalar

The margin surrounding this box.

Note that unlike most other geometries, a box has an interior margin. Therefore, the real extents of the box (those that have been passed to the Cuboid constructor) equal the sum of the margin and its half-extents (as returned by the half_extents method).