Struct ncollide2df32​::geom​::Mesh​

pub struct Mesh {
    // some fields omitted
}

Geometry commonly known as a 2d line strip or a 3d triangle mesh.

Methods

impl Mesh

fn new(vertices: Arc<Vec<Vect>>, indices: Arc<Vec<uint>>, uvs: Option<Arc<Vec<Vec2<Scalar>>>>, normals: Option<Arc<Vec<Vect>>>) -> Mesh

Builds a new mesh with a default margin of 0.04.

fn new_with_margin(vertices: Arc<Vec<Vect>>, indices: Arc<Vec<uint>>, uvs: Option<Arc<Vec<Vec2<Scalar>>>>, normals: Option<Arc<Vec<Vect>>>, margin: Scalar) -> Mesh

Builds a new mesh with a custom margin.

impl Mesh

fn vertices<'a>(&'a self) -> &'a Arc<Vec<Vect>>

The vertices of this mesh.

fn bounding_volumes<'a>(&'a self) -> &'a [AABB]

Bounding volumes of the subsimplices.

fn indices<'a>(&'a self) -> &'a Arc<Vec<uint>>

The indices of this mesh.

fn uvs<'a>(&'a self) -> &'a Option<Arc<Vec<Vec2<Scalar>>>>

The texture coordinates of this mesh.

fn normals<'a>(&'a self) -> &'a Option<Arc<Vec<Vect>>>

The normals of this mesh.

fn bvt<'a>(&'a self) -> &'a BVT<uint, AABB>

The acceleration structure used for efficient collision detection and ray casting.

fn margin(&self) -> Scalar

The collision margin used by this mesh.

impl Mesh

fn element_at(&self, i: uint) -> MeshPrimitive

Gets the i-th mesh element.