Trait ncollide2df32​::geom​::ConcaveGeom​

pub trait ConcaveGeom: Geom {
    fn map_part_at<T>(&self, uint, |&Matrix, &Geom| -> T) -> T;
    fn map_transformed_part_at<T>(&self, m: &Matrix, uint, |&Matrix, &Geom| -> T) -> T;
    fn approx_interferences_with_aabb(&self, &AABB, &mut Vec<uint>);
    fn approx_interferences_with_ray(&self, &Ray, &mut Vec<uint>);
    fn aabb_at<'a>(&'a self, i: uint) -> &'a AABB;
}

Trait implemented by concave, composite geometries.

A composite geometry is composed of several Geom. Typically, it is a convex decomposition of a concave geometry.

Required Methods

fn map_part_at<T>(&self, uint, |&Matrix, &Geom| -> T) -> T

Applies a function to each sub-geometry of this concave geometry.

fn map_transformed_part_at<T>(&self, m: &Matrix, uint, |&Matrix, &Geom| -> T) -> T

Applies a transformation matrix and a function to each sub-geometry of this concave geometry.

fn approx_interferences_with_aabb(&self, &AABB, &mut Vec<uint>)

Computes the indices of every sub-geometry which might intersect a given AABB.

fn approx_interferences_with_ray(&self, &Ray, &mut Vec<uint>)

Computes the indices of every sub-geometry which might intersect a given Ray.

fn aabb_at<'a>(&'a self, i: uint) -> &'a AABB

Gets the AABB of the geometry identified by the index i.