Struct ncollide2df32​::narrow​::GeomConcaveGeom​

pub struct GeomConcaveGeom<G1, G2> {
    // some fields omitted
}

Collision detector between a geometry and a concave geometry.

Methods

impl<G1, G2> GeomConcaveGeom<G1, G2>

fn new() -> GeomConcaveGeom<G1, G2>

Creates a new collision detector between a geometry and a concave geometry.