Struct ncollide2df32​::narrow​::GeomGeomDispatcher​

pub struct GeomGeomDispatcher {
    // some fields omitted
}

Collision dispatcher between two ~Geom.

Methods

impl GeomGeomDispatcher

fn new_without_default() -> GeomGeomDispatcher

Creates a new GeomGeomDispatcher without the default set of collision detectors factories.

unsafe fn register_factory<G1: 'static + Any, G2: 'static + Any, F: CollisionDetectorFactory>(&mut self, factory: F)

Registers a new collision detection algorithm factory for a pair of geometries.

This is unsafe because there is no way to check that the factory will really generate collision detectors suited for G1 and G2. Whenever possible, use register_detector or register_dynamic_detector instead.

fn register_dynamic_detector<G1: 'static + Any, G2: 'static + Any, D: 'static + Send + Clone + DynamicCollisionDetector<G1, G2>>(&mut self, d: D)

Registers a new dynamic collision detector for two geometries.

fn register_detector<G1: 'static + Any, G2: 'static + Any, D: 'static + Send + CollisionDetector<G1, G2> + Clone>(&mut self, d: D)

Registers a new collision detector for two geometries.

fn unregister_detector<G1: 'static + Any, G2: 'static + Any>(&mut self)

Unregister the collision detector for a givem pair of geometries.

fn dispatch(&self, a: &Geom, b: &Geom) -> Option<Box<GeomGeomCollisionDetector>>

If registered, creates a new collision detector adapted for the two given geometries.

impl GeomGeomDispatcher

fn new() -> GeomGeomDispatcher

Creates a new GeomGeomDispatcher able do build collision detectors for any valid pair of geometries supported by ncollide.

fn register_default_plane_implicit_detector<I: 'static + Implicit<Vect, Matrix>>(&mut self, generate_manifold: bool, prediction: &Scalar)

Registers a PlaneImplicit collision detector between a given implicit geometry and a plane.

fn register_default_implicit_implicit_detector<G1: 'static + Implicit<Vect, Matrix> + PreferedSamplingDirections<Vect, Matrix>, G2: 'static + Implicit<Vect, Matrix> + PreferedSamplingDirections<Vect, Matrix>, S: Send + Clone + Simplex<AnnotatedPoint>>(&mut self, generate_manifold: bool, prediction: &Scalar, simplex: &S)

Register an ImplicitImplicit collision detector between two implicit geometries.

fn register_default_concave_geom_geom_detector<G1: 'static + ConcaveGeom, G2: 'static + Geom>(&mut self)

Register an ConcaveGeomGeom collision detector between a given concave geometry and a given geometry.

fn register_detector_with_contact_manifold_generator<G1: 'static + Any, G2: 'static + Any, D: 'static + Send + CollisionDetector<G1, G2> + Clone>(&mut self, d: D, prediction: &Scalar)

Register a given collision detector and adds it a contact manifold generator (a OneShotContactManifoldGenerator).

fn register_default_implicit_detectors<G: 'static + Implicit<Vect, Matrix> + PreferedSamplingDirections<Vect, Matrix>>(&mut self, generate_manifold: bool, prediction: &Scalar)

Register ImplicitImplicit collision detectors between a given geometry and every implicit geometry supported by ncollide.