Struct ncollide2df32​::narrow​::surface_subdivision_tree​::SurfaceSubdivisionTreeCache​

pub struct SurfaceSubdivisionTreeCache<D> {
    // some fields omitted
}

A cache that keeps track of parametric surface subdivision trees.

This cache allows only insersion. Deletion is automatic.

Methods

impl<D: Send + Sync> SurfaceSubdivisionTreeCache<D>

fn new() -> SurfaceSubdivisionTreeCache<D>

Creates a new surface subdivision tree cache.

fn clear(&mut self)

Removes everything from this cache.

fn find_or_insert_with(cache: &mut Arc<RWLock<SurfaceSubdivisionTreeCache<D>>>, b: &BezierSurface, data: || -> D) -> SurfaceSubdivisionTreeRef<D>

Gets from the cache cache, the subdivision tree for the surface b.