Module ncollide2df32​::partitioning​

Spatial partitioning tools.

Reexports

pub use partitioning::bvt::{Internal, Leaf};

Structs

BVT
A Boundig Volume Tree.

BoundingVolumeInterferencesCollector
Bounding Volume Tree visitor collecting interferences with a given bounding volume.

DBVT
A Dynamic Bounding Volume Tree.

DBVTLeaf
Leaf of a Dynamic Bounding Volume Tree.

RayInterferencesCollector
Bounding Volume Tree visitor collecting interferences with a given ray.

Enums

BVTNode
A node of the bounding volume tree.

BinaryPartition
Result of a binary partition.

DBVTLeafState
State of a leaf.

Traits

BVTTVisitor
Visitor for the Bounding Volume Traversal Tree.

BVTVisitor
Visitor of Bounding Volume Trees.

Functions

kdtree_partitioner
Construction function for a kdree to be used with BVT::new_with_partitioner.

kdtree_partitioner_with_centers
Construction function for a kdree to be used with BVT::new_with_partitioner.