Enum ncollide2df32​::partitioning​::DBVTLeafState​

pub enum DBVTLeafState<B, BV> {
    RightChildOf(*mut DBVTInternal<B, BV>),
    LeftChildOf(*mut DBVTInternal<B, BV>),
    Detached,
}

State of a leaf.

Variants

RightChildOf

This leaf is the right child of another node.

LeftChildOf

This leaf is the left child of another node.

Detached

This leaf is detached from any tree.

Methods

impl<B, BV> DBVTLeafState<B, BV>

fn is_detached(&self) -> bool

Indicates whether this leaf is detached.