Enum ncollide2df32​::procedural​::path​::PathSample​

pub enum PathSample<V> {
    StartPoint(V, V),
    InnerPoint(V, V),
    EndPoint(V, V),
    EndOfSample,
}

A sample point and its associated tangent.

Variants

StartPoint

A point that starts a new path.

InnerPoint

A point that is inside of the path currently generated.

EndPoint

A point that ends the path currently generated.

EndOfSample

Used when the sampler does not have any other points to generate.