Trait ncollide2df32​::volumetric​::volumetric​::Volumetric​

pub trait Volumetric {
    fn surface(&self) -> Scalar;
    fn volume(&self) -> Scalar;
    fn center_of_mass(&self) -> Vect;
    fn unit_angular_inertia(&self) -> AngularInertia;

    fn mass(&self, density: &Scalar) -> Scalar { ... }
    fn angular_inertia(&self, mass: &Scalar) -> AngularInertia { ... }
    fn mass_properties(&self, density: &Scalar) -> (Scalar, Vect, AngularInertia) { ... }
}

Trait to be implemented by objects which have a mass, a center of mass, and an inertia tensor.

Required Methods

fn surface(&self) -> Scalar

Computes the surface of this object.

fn volume(&self) -> Scalar

Computes the volume of this object.

fn center_of_mass(&self) -> Vect

Computes the center of mass of this object.

fn unit_angular_inertia(&self) -> AngularInertia

Computes the angular inertia tensor of this object.

Provided Methods

fn mass(&self, density: &Scalar) -> Scalar

Given its density, this computes the mass of this object.

fn angular_inertia(&self, mass: &Scalar) -> AngularInertia

Given its mass, this computes the angular inertia of this object.

fn mass_properties(&self, density: &Scalar) -> (Scalar, Vect, AngularInertia)

Given its density, this computes the mass, center of mass, and inertia tensor of this object.